Address RNA read phasing review
