+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 **Added**
 - Added `RNAReadPhasingSource`, a BAM-backed `ReadPhasingSource`
   implementation for RNA read/fragment co-occurrence. It is consumed
-  through the existing `ReadPhaseResolver(source)` API and lives behind
-  the optional `varcode[rna]` / `pysam` dependency.
+  through `MolecularPhaseResolver(source)` and lives behind the optional
+  `varcode[rna]` / `pysam` dependency. `ReadPhaseResolver` remains as
+  the varcode 5.0 compatibility name.
 
 **Breaking**
 - Unified the multi-outcome machinery: `SpliceCandidate` deleted;
