+++ docs/api.md
@@ -117,6 +117,10 @@ Auto-generated from in-source docstrings via
 
 ::: varcode.MutantTranscriptSource
 
+### `varcode.MolecularPhaseResolver`
+
+::: varcode.MolecularPhaseResolver
+
 ### `varcode.ReadPhaseResolver`
 
 ::: varcode.ReadPhaseResolver
