+++ docs/germline.md
@@ -21,13 +21,14 @@ possibility set** — one classified effect per haplotype hypothesis
 
 - `VCFPhaseResolver(merged_phased.vcf)` — reads `PS` tags from a
   WhatsHap- or HapCUT2-phased merged VCF.
-- `ReadPhaseResolver(source)` — wraps any RNA-phasing source
+- `MolecularPhaseResolver(source)` — wraps any RNA-phasing source
   (typically an Isovar adapter shipped by `openvax/isovar`) to check
   which haplotype the somatic was observed on in RNA reads.
   For RNA-seq BAMs without assembly, use
-  `ReadPhaseResolver(RNAReadPhasingSource("tumor.rna.bam"))` to phase
+  `MolecularPhaseResolver(RNAReadPhasingSource("tumor.rna.bam"))` to phase
   by direct read/fragment co-occurrence. Raw BAM phasing does not
   provide observed `MutantTranscript`s; assembly-backed sources can.
+  `ReadPhaseResolver` remains as the varcode 5.0 compatibility name.
 - Anything implementing `in_cis(v1, v2, transcript) -> bool | None`.
 
 Phase known → single `MutationEffect`. Phase unknown →
@@ -115,7 +116,7 @@ print("trans ->", type(eff_trans).__name__, eff_trans.short_description)
 ```
 
 In real code use `VCFPhaseResolver(merged_phased.vcf)` or
-`ReadPhaseResolver(source)` — same `phase_resolver=` slot.
+`MolecularPhaseResolver(source)` — same `phase_resolver=` slot.
 
 ## When does this matter?
